+++ python/src/models/deep/xlstm.py
@@ -4,7 +4,7 @@ xLSTM Model wrapping xLSTMBlocks.
 
 from torch import nn
 
-from ..modules.xlstm_block import xLSTMBlock
+from .modules.xlstm_block import xLSTMBlock
 
 
 class xLSTM(nn.Module):