+++ python/src/models/helper/association_rule.py
@@ -6,6 +6,7 @@ import torch
 from ..mac.base import ClassicalModel
 from .logic_association.apriori import AprioriAlgorithm
 from .logic_association.fpgrowth import FPGrowthAlgorithm
+from .logic_association.eclat import EclatAlgorithm
 
 class AssociationRuleModel(ClassicalModel):
     """Base class for association rule models."""
@@ -34,3 +35,8 @@ class FPGrowthModel(AssociationRuleModel):
     def __init__(self, min_support=0.5, min_confidence=0.7, **kwargs):
         super().__init__()
         self.model = FPGrowthAlgorithm(min_support=min_support, min_confidence=min_confidence, **kwargs)
+
+class EclatModel(AssociationRuleModel):
+    def __init__(self, min_support=0.5, min_confidence=0.7, **kwargs):
+        super().__init__()
+        self.model = EclatAlgorithm(min_support=min_support, min_confidence=min_confidence, **kwargs)