+++ python/src/models/helper/clustering.py
@@ -7,6 +7,7 @@ from .logic_clustering.kmeans import KMeansAlgorithm
 from .logic_clustering.hierarchical import HierarchicalClusteringAlgorithm
 from .logic_clustering.dbscan import DBSCANAlgorithm
 from .logic_clustering.gmm import GMMAlgorithm
+from .logic_clustering.kmedians import KMediansAlgorithm
 
 class ClusteringModel(ClassicalModel):
     """Base class for clustering models."""
@@ -32,3 +33,12 @@ class GMMModel(ClusteringModel):
     def __init__(self, n_components=1, **kwargs):
         super().__init__()
         self.model = GMMAlgorithm(n_components=n_components, **kwargs)
+
+class EMModel(GMMModel):
+    """Expectation Maximisation (EM) clustering. Alias for Gaussian Mixture Model."""
+    pass
+
+class KMediansModel(ClusteringModel):
+    def __init__(self, n_clusters=8, **kwargs):
+        super().__init__()
+        self.model = KMediansAlgorithm(n_clusters=n_clusters, **kwargs)